+++ docs/changes.rst
@@ -46,6 +46,15 @@ v1.2.0 (UNRELEASED)
   message got a reply, potentially making the application wait forever on
   replies to the subsequent stop messages.
 
+- When :meth:`~pykka.ActorRef.ask` is used to asynchronously message a dead
+  actor (e.g. ``block`` set to :class:`False`), it will no longer immediately
+  raise :exc:`~pykka.ActorDeadError`. Instead, it will return a future and
+  fail the future with the :exc:`~pykka.ActorDeadError` exception. This makes
+  the interace more consistent, as you'll have one instead of two ways the call
+  can raise exceptions under normal conditions. If :meth:`~pykka.ActorRef.ask`
+  is called synchronously (e.g. ``block`` set to :class:`True`), the behavior
+  is unchanged.
+
 
 v1.1.0 (2013-01-19)
 ===================